+++ string/cpp_string/cpp_string.hpp
@@ -12,12 +12,16 @@ extern "C" {
 	cpp_object* cpp_string_clone( cpp_object* );
 	cpp_object* cpp_string_int2str( int num);
 	cpp_object* cpp_string_double2str( double num);
+	const char* cpp_string_int2cstr(int num);
+	const char* cpp_string_double2cstr(double num);
 	cpp_object* cpp_string_lstrip( cpp_object* );
 	cpp_object* cpp_string_rstrip( cpp_object* );
 	cpp_object* cpp_string_strip( cpp_object* );
 	const char* cpp_string_read (cpp_object* );
 	cpp_object* cpp_string_concat (cpp_object*, cpp_object*);
 	cpp_object** cpp_string_ptr_concat (cpp_object*, cpp_object*);
+	void cpp_string_append_string (cpp_object*, cpp_object*);
+	void cpp_string_append_cstring (cpp_object*, const char*);
 	cpp_object* cpp_string_repeat (cpp_object*, int);
 	cpp_object* cpp_string_subset (cpp_object*, size_t, size_t , const char* encoding);
 	int cpp_string_has_char (cpp_object*, char);