+++ test/test_func.h
@@ -15,18 +15,18 @@ test_func_test1( void )
 {
 	// Code 
 	const char* code = " "
-"space_pi = ' ' + str_subset(num_to_str(3.14),0,3) + num_to_str(1592) + num_to_str(6535) ;"
+"space_pi = ' ' + str_subset(num_to_str(3.14),1,4) + num_to_str(1592) + num_to_str(6535) ;"
 "pi_str = 'PI is' + space_pi ;"
-"print(pi_str);"
 "pi_num = str_to_num( str_strip(space_pi)) ;"
-"three_str = str_subset( num_to_str(10/3) , 0 , 3);"
+"three_str = str_subset( num_to_str(10/3) , 1 , 4);"
 "seven_str = str_repeat( num_to_str(7), 7);"
 "ten_str = num_to_str( 2.5 * 4 ) \n"
 "\n"
 "spaced_str = '    hello world    ';"
-"print(spaced_str);"
 "stripped_str = str_strip(spaced_str);"
-"print(stripped_str);"
+"jpn_hello = 'こんにちは';"
+"new_str = str_concat( jpn_hello, ' ' , stripped_str );"
+"new_str2 = str_concat( new_str, ', PI = ', pi_num);"
 ;
 
 	// Parser Initialization
@@ -45,6 +45,10 @@ test_func_test1( void )
 
 	sailr_ptr_table_create_null(&table, "spaced_str");
 	sailr_ptr_table_create_null(&table, "stripped_str");
+
+	sailr_ptr_table_create_null(&table, "jpn_hello");
+	sailr_ptr_table_create_null(&table, "new_str");
+	sailr_ptr_table_create_null(&table, "new_str2");
 
 	// Creating virtual machine codes
 	vm_inst_object* inst_list = sailr_gen_code( ps, table); // VM Code is generated.
@@ -66,6 +70,8 @@ test_func_test1( void )
 	char st_ten_str = sailr_ptr_table_get_type(&table, "ten_str");
 	char st_spaced_str = sailr_ptr_table_get_type(&table, "spaced_str");
 	char st_stripped_str = sailr_ptr_table_get_type(&table, "stripped_str");
+	char st_new_str = sailr_ptr_table_get_type(&table, "new_str");
+	char st_new_str2 = sailr_ptr_table_get_type(&table, "new_str2");
 
 	CU_ASSERT_EQUAL( st_space_pi , 's');
 	CU_ASSERT_EQUAL( st_pi_str , 's');
@@ -75,6 +81,8 @@ test_func_test1( void )
 	CU_ASSERT_EQUAL( st_ten_str , 's');
 	CU_ASSERT_EQUAL( st_spaced_str , 's');
 	CU_ASSERT_EQUAL( st_stripped_str , 's');
+	CU_ASSERT_EQUAL( st_new_str , 's');
+	CU_ASSERT_EQUAL( st_new_str2 , 's');
 
 
 	const char* s_space_pi = sailr_ptr_table_read_string(&table, "space_pi");
@@ -85,6 +93,8 @@ test_func_test1( void )
 	const char* s_ten_str = sailr_ptr_table_read_string(&table, "ten_str");
 	const char* s_spaced_str = sailr_ptr_table_read_string(&table, "spaced_str");
 	const char* s_stripped_str = sailr_ptr_table_read_string(&table, "stripped_str");
+	const char* s_new_str = sailr_ptr_table_read_string(&table, "new_str");
+	const char* s_new_str2 = sailr_ptr_table_read_string(&table, "new_str2");
 
 	CU_ASSERT_STRING_EQUAL( s_space_pi , " 3.1415926535");
 	CU_ASSERT_STRING_EQUAL( s_pi_str , "PI is 3.1415926535");
@@ -94,6 +104,8 @@ test_func_test1( void )
 	CU_ASSERT_STRING_EQUAL( s_ten_str , "10.000000");
 	CU_ASSERT_STRING_EQUAL( s_spaced_str , "    hello world    ");
 	CU_ASSERT_STRING_EQUAL( s_stripped_str , "hello world");
+	CU_ASSERT_STRING_EQUAL( s_new_str , "こんにちは hello world");
+	CU_ASSERT_STRING_EQUAL( s_new_str2 , "こんにちは hello world, PI = 3.141593" );
 
 	// Clean up
 	sailr_tree_free(ps);